Refactor Query History to floating panel; remove favorites dropdown
Agent-Logs-Url: https://github.com/FalkorDB/falkordb-browser/sessions/cebb789b-4445-40ad-800a-947516562ca7
